Backend/db/supabase_client.py
import os
from supabase import create_client
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")

if not SUPABASE_URL or not SUPABASE_KEY:
    raise Exception("Supabase credentials are missing in .env file")

# Create Supabase client
supabase = create_client(SUPABASE_URL, SUPABASE_KEY)

def get_supabase():
    """
    Returns the Supabase client.
    Import this function wherever DB access is needed.
    """
    return supabase
