Backend/requirements.txt
flask
flask-cors
python-dotenv
supabase
flask-jwt-extended
passlib[bcrypt]
pydantic
